updated GA publish timing events
